# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ComfyUI Studio API — proxy and orchestration layer for a ComfyUI server
servers:
  - url: /api
    description: Base API path
security:
  - clerkSession: []
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get current server settings
      responses:
        "200":
          description: Settings object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update server settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
...
  /settings/gpus:
    post:
      operationId: addSavedGpu
      tags: [settings]
      summary: Save a GPU endpoint URL under a label for one-click switching
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedGpuInput"
      responses:
        "201":
          description: Saved GPU endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedGpu"
        "400":
          description: Invalid request
  /settings/gpus/{id}:
    delete:
      operationId: deleteSavedGpu
      tags: [settings]
      summary: Remove a saved GPU endpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
        "404":
          description: Not found
...
  /saved-workflows:
    get:
      operationId: listSavedWorkflows
      tags: [saved-workflows]
      summary: List user-saved custom workflows
      responses:
        "200":
          description: Saved workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedWorkflow"
    post:
      operationId: createSavedWorkflow
      tags: [saved-workflows]
      summary: Save a custom workflow by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedWorkflowInput"
      responses:
        "201":
          description: Saved workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedWorkflow"
        "400":
          description: Invalid request
  /saved-workflows/{id}:
    delete:
      operationId: deleteSavedWorkflow
      tags: [saved-workflows]
      summary: Delete a saved workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
        "404":
          description: Not found
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk browser session cookie. Sign in through the web client before calling protected endpoints.
...
  schemas:
...
    Settings:
      type: object
      required: [comfyUrl, updatedAt, savedGpus, routingMode, selectedGpuId]
      properties:
        comfyUrl:
          type: string
          description: URL of the ComfyUI server (e.g. http://localhost:8188)
        updatedAt:
          type: string
          format: date-time
        savedGpus:
          type: array
          items:
            $ref: "#/components/schemas/SavedGpu"
        routingMode:
          type: string
          enum: [auto, manual]
        selectedGpuId:
          type: ["number", "null"]
    SavedGpu:
      type: object
      required: [id, label, url]
      properties:
        id:
          type: number
        label:
          type: string
        url:
          type: string
    SavedGpuInput:
      type: object
      required: [label, url]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 50
        url:
          type: string
          minLength: 1
    SettingsInput:
      type: object
      required: [comfyUrl]
      properties:
        comfyUrl:
          type: string
        routingMode:
          type: string
          enum: [auto, manual]
        selectedGpuId:
          type: ["number", "null"]
...
    SavedWorkflow:
      type: object
      required: [id, name, json, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        json:
          type: string
          description: The ComfyUI API-format workflow JSON as a string
        createdAt:
          type: string
    SavedWorkflowInput:
      type: object
      required: [name, json]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        json:
          type: string
          minLength: 2
...